+++ django_enum/tests/tests.py
@@ -16,6 +16,7 @@ from django.utils.functional import classproperty
 from django_enum import TextChoices
 from django_enum.choices import choices, labels, names, values
 from django_enum.forms import EnumChoiceField  # dont remove this
+from django.forms import Form, ModelForm
 # from django_enum.tests.djenum.enums import (
 #     BigIntEnum,
 #     BigPosIntEnum,
@@ -413,6 +414,28 @@ class TestChoices(EnumTypeMixin, TestCase):
                 1
             )
         self.MODEL_CLASS.objects.all().delete()
+
+    def test_coerce_to_primitive(self):
+
+        create_params = {
+            **self.create_params,
+            'no_coerce': '32767'
+        }
+
+        tester = self.MODEL_CLASS.objects.create(**create_params)
+
+        self.assertIsInstance(tester.no_coerce, int)
+        self.assertEqual(tester.no_coerce, 32767)
+
+    def test_coerce_to_primitive_error(self):
+
+        create_params = {
+            **self.create_params,
+            'no_coerce': 'Value 32767'
+        }
+
+        with self.assertRaises(ValueError):
+            self.MODEL_CLASS.objects.create(**create_params)
 
     def test_to_python_deferred_attribute(self):
         obj = self.MODEL_CLASS.objects.create(**self.create_params)
@@ -1813,6 +1836,109 @@ class TestBulkOperations(EnumTypeMixin, TestCase):
         )
 
 
+class FormTests(EnumTypeMixin, TestCase):
+    """
+    Some more explicit form tests that allow easier access to other internal workflows.
+    """
+
+    MODEL_CLASS = EnumTester
+
+    @property
+    def model_form_class(self):
+
+        class EnumTesterForm(ModelForm):
+
+            class Meta:
+                model = self.MODEL_CLASS
+                fields = '__all__'
+
+        return EnumTesterForm
+
+    @property
+    def basic_form_class(self):
+        from django.core.validators import MinValueValidator, MaxValueValidator
+
+        class BasicForm(Form):
+            
+            small_pos_int = EnumChoiceField(self.SmallPosIntEnum)
+            small_int = EnumChoiceField(self.SmallIntEnum)
+            pos_int = EnumChoiceField(self.PosIntEnum)
+            int = EnumChoiceField(self.IntEnum)
+            big_pos_int = EnumChoiceField(self.BigPosIntEnum)
+            big_int = EnumChoiceField(self.BigIntEnum)
+            constant = EnumChoiceField(self.Constants)
+            text = EnumChoiceField(self.TextEnum)
+            extern = EnumChoiceField(self.ExternEnum)
+            dj_int_enum = EnumChoiceField(self.DJIntEnum)
+            dj_text_enum = EnumChoiceField(self.DJTextEnum)
+            non_strict_int = EnumChoiceField(self.SmallPosIntEnum, strict=False)
+            non_strict_text = EnumChoiceField(self.TextEnum, strict=False)
+            no_coerce = EnumChoiceField(
+                self.SmallPosIntEnum,
+                validators=[MinValueValidator(0), MaxValueValidator(32767)]
+            )
+        
+        return BasicForm
+
+    @property
+    def test_params(self):
+        return {
+            'small_pos_int': self.SmallPosIntEnum.VAL2,
+            'small_int': self.SmallIntEnum.VALn1,
+            'pos_int': self.PosIntEnum.VAL3,
+            'int': self.IntEnum.VALn1,
+            'big_pos_int': self.BigPosIntEnum.VAL3,
+            'big_int': self.BigIntEnum.VAL2,
+            'constant': self.Constants.GOLDEN_RATIO,
+            'text': self.TextEnum.VALUE2,
+            'extern': self.ExternEnum.TWO,
+            'dj_int_enum': self.DJIntEnum.THREE,
+            'dj_text_enum': self.DJTextEnum.A,
+            'non_strict_int': '15',
+            'non_strict_text': 'arbitrary',
+            'no_coerce': self.SmallPosIntEnum.VAL3
+        }
+    
+    @property
+    def test_data_strings(self):
+        return {key: str(value) for key, value in self.test_params.items()}
+
+    @property
+    def expected(self):
+        return {
+            **self.test_params,
+            'non_strict_int': int(self.test_params['non_strict_int']),
+        }
+
+    def test_modelform_binding(self):
+        form = self.model_form_class(data=self.test_data_strings)
+
+        form.full_clean()
+        self.assertTrue(form.is_valid())
+
+        for key, value in self.expected.items():
+            self.assertEqual(form.cleaned_data[key], value)
+
+        self.assertIsInstance(form.cleaned_data['no_coerce'], int)
+        self.assertIsInstance(form.cleaned_data['non_strict_int'], int)
+
+        obj = form.save()
+
+        for key, value in self.expected.items():
+            self.assertEqual(getattr(obj, key), value)
+
+    def test_basicform_binding(self):
+        form = self.basic_form_class(data=self.test_data_strings)
+        form.full_clean()
+        self.assertTrue(form.is_valid())
+
+        for key, value in self.expected.items():
+            self.assertEqual(form.cleaned_data[key], value)
+
+        self.assertIsInstance(form.cleaned_data['no_coerce'], int)
+        self.assertIsInstance(form.cleaned_data['non_strict_int'], int)
+
+
 if ENUM_PROPERTIES_INSTALLED:
 
     from django_enum.forms import EnumChoiceField
@@ -1843,6 +1969,9 @@ if ENUM_PROPERTIES_INSTALLED:
     )
     from enum_properties import EnumProperties, s
 
+    class EnumPropertiesFormTests(FormTests):
+
+        MODEL_CLASS = EnumTester
 
     class TestEnumPropertiesIntegration(TestCase):
 
@@ -3362,6 +3491,22 @@ if ENUM_PROPERTIES_INSTALLED:
             self.assertTrue(tester._meta.get_field('dj_text_enum').validate('A', tester) is None)
             self.assertTrue(tester._meta.get_field('non_strict_int').validate(20, tester) is None)
 
+        def test_coerce_to_primitive_error(self):
+            """
+            Override this base class test because this should work with symmetrical enum.
+            """
+            create_params = {
+                **self.create_params,
+                'no_coerce': 'Value 32767'
+            }
+
+            tester = self.MODEL_CLASS.objects.create(**create_params)
+            self.assertEqual(tester.no_coerce, self.SmallPosIntEnum.VAL3)
+            self.assertEqual(tester.no_coerce, 'Value 32767')
+
+            tester.refresh_from_db()
+            self.assertEqual(tester.no_coerce, 32767)
+            
     class PerformanceTest(TestCase):
         """
         We intentionally test bulk operations performance because thats what
@@ -3660,7 +3805,6 @@ if ENUM_PROPERTIES_INSTALLED:
                 ).first() == instance
             )
 
-            from django.forms import ModelForm
             from django_enum import EnumChoiceField
 
             class TextChoicesExampleForm(ModelForm):
@@ -3728,4 +3872,3 @@ if ENUM_PROPERTIES_INSTALLED:
 
 else:  # pragma: no cover
     pass
-