+++ doc/source/changelog.rst
@@ -2,10 +2,12 @@
 Change Log
 ==========
 
-v1.2.3
+v1.3.0
 ======
 
-* Added `Support Django 5.0 <https://github.com/bckohan/django-enum/issues/54>`_
+* Implemented `Support db_default <https://github.com/bckohan/django-enum/issues/56>`_
+* Fixed `When coerce=False, enum form fields and model fields should still coerce to the enum's primitive type. <https://github.com/bckohan/django-enum/issues/55>`_
+* Implemented `Support Django 5.0 <https://github.com/bckohan/django-enum/issues/54>`_
 
 v1.2.2
 ======